+++ src/r_alias.c
@@ -664,7 +664,7 @@ void R_AliasDrawModel(alight_t* plighting)
     r_amodels_drawn++;
 
     // cache align
-    pfinalverts = (finalvert_t*)(((long)&finalverts[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+    pfinalverts = (finalvert_t*)(((size_t)&finalverts[0] + CACHE_SIZE - 1) & ~(size_t)(CACHE_SIZE - 1));
     pauxverts = &auxverts[0];
 
     paliashdr = (aliashdr_t*)Mod_Extradata(currententity->model);
